requirements.txt
Pillow
ruff
pre-commit
pandas
scikit-learn
seaborn
matplotlib
pytest
pytest-cov